Redesign font BG to require set_bg_padding()
It's kind of odd, but fine for now.

The previous design subtracted the padding from the (x,y) to draw the
background color at a negative offset, which wasn't very intuitive
and required remembering to add the padding to the FG text pos.

Now, you just need to make sure to call set_bg_padding() before calling
any other func, but don't need to worry about changing the pos manually.

However, one caveat is that you should now add to the pos in the cases
where you do manually alter the pos, instead of overwriting it.